refactor(package): 정확하지 않은 패키지 명명 수정
