+++ src/main/java/lotto/model/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.model.vo;
+package lotto.model.domain;
 
 import java.util.ArrayList;
 import java.util.List;
