+++ src/main/java/lotto/model/domain/WinCriterion.java
@@ -1,4 +1,4 @@
-package lotto.model.vo;
+package lotto.model.domain;
 
 import lotto.model.constants.ErrorMessage;
