+++ src/main/java/lotto/model/dto/LottoResult.java
@@ -1,7 +1,7 @@
 package lotto.model.dto;
 
-import lotto.model.vo.Profit;
-import lotto.model.vo.WinResult;
+import lotto.model.domain.Profit;
+import lotto.model.domain.WinResult;
 
 public class LottoResult {
