+++ src/main/java/lotto/service/LottoService.java
@@ -1,12 +1,12 @@
 package lotto.service;
 
 import lotto.model.dto.PurchaseAmount;
-import lotto.model.vo.Lotto;
-import lotto.model.vo.Lottos;
-import lotto.model.vo.Profit;
-import lotto.model.vo.PurchaseInfo;
-import lotto.model.vo.WinLotto;
-import lotto.model.vo.WinResult;
+import lotto.model.domain.Lotto;
+import lotto.model.domain.Lottos;
+import lotto.model.domain.Profit;
+import lotto.model.domain.PurchaseInfo;
+import lotto.model.domain.WinLotto;
+import lotto.model.domain.WinResult;
 
 public interface LottoService {
