+++ src/main/java/lotto/service/LottoServiceImpl.java
@@ -3,14 +3,14 @@ package lotto.service;
 import java.util.Collections;
 import lotto.model.constants.LottoConstants;
 import lotto.model.dto.PurchaseAmount;
-import lotto.model.vo.Lotto;
-import lotto.model.vo.LottoNumber;
-import lotto.model.vo.Lottos;
-import lotto.model.vo.MatchCounts;
-import lotto.model.vo.Profit;
-import lotto.model.vo.PurchaseInfo;
-import lotto.model.vo.WinLotto;
-import lotto.model.vo.WinResult;
+import lotto.model.domain.Lotto;
+import lotto.model.domain.LottoNumber;
+import lotto.model.domain.Lottos;
+import lotto.model.domain.MatchCounts;
+import lotto.model.domain.Profit;
+import lotto.model.domain.PurchaseInfo;
+import lotto.model.domain.WinLotto;
+import lotto.model.domain.WinResult;
 
 public class LottoServiceImpl implements LottoService {
