+++ src/main/java/lotto/view/ResultView.java
@@ -1,10 +1,10 @@
 package lotto.view;
 
 import lotto.model.dto.LottoResult;
-import lotto.model.vo.Lottos;
-import lotto.model.vo.Profit;
-import lotto.model.vo.PurchaseInfo;
-import lotto.model.vo.WinResult;
+import lotto.model.domain.Lottos;
+import lotto.model.domain.Profit;
+import lotto.model.domain.PurchaseInfo;
+import lotto.model.domain.WinResult;
 
 public class ResultView {
