+++ src/test/java/lotto/model/domain/PurchaseInfoTest.java
@@ -1,4 +1,4 @@
-package lotto.model.vo;
+package lotto.model.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
